Insertando control de errores en el controlador

+++ src/main/java/controlador/Submenus.java
@@ -24,22 +24,34 @@ public class Submenus {
 				case 2:
 					lee.Print("Introduzca el nombre del producto a eliminar");
 					s=lee.String();
-					productos.deleteProducto(s);
+					if(productos.searchProducto(s)!=null) {
+						productos.deleteProducto(s);
+						lee.Print("producto "+s+" eliminado");
+					}else {
+						lee.Print("No existe el producto: "+s);
+					}
 					break;
 				case 3:
 					lee.Print("Introduzca el nombre del producto que quiera modificar");
 					s=lee.String();
-					Editmenu.Producto(s, productos);
+					if(productos.searchProducto(s)!=null) {
+						Editmenu.Producto(s, productos);
+					}else {
+						lee.Print("No existe el producto: "+s);
+					}
 					break;
 				case 4:
 					lee.Print("Introduzca el nombre del producto que quiera buscar");
 					s=lee.String();
-					Producto p = new Producto();
-					p=productos.searchProducto(s);
-					lee.Print(p);
+					if(productos.searchProducto(s)!=null) {
+						Producto p = new Producto();
+						p=productos.searchProducto(s);
+						lee.Print(p);
+					}else {
+						lee.Print("No existe el producto: "+s);
+					}
 					break;
 				case 5:
-					
 					lee.show(productos);
 					break;
 				case 6:
@@ -71,23 +83,38 @@ public class Submenus {
 				case 2:
 					lee.Print("Introduzca el dni del cliente a eliminar");
 					s=lee.String();
-					try {
-					clientes.deleteCliente(s);
-					}catch(Exception e) {
-						lee.Print("No se ha podido elimar al cliente o no existe");
+					if(clientes.searchCliente(s)!=null) {
+						try {
+							clientes.deleteCliente(s);
+							}catch(Exception e) {
+								lee.Print("No se ha podido elimar al cliente o no existe");
+							}
+						lee.Print("Cliente "+s+" eliminado");
+					}else {
+						lee.Print("No se ha podido eliminar al cliente "+s+" o no existe");
 					}
+
 					break;
 				case 3:
 					lee.Print("Introduzca el dni del cliente que quiera modificar");
 					s=lee.String();
-					Editmenu.Cliente(s, clientes);
+					if(clientes.searchCliente(s)!=null) {
+						Editmenu.Cliente(s, clientes);
+					}else {
+						lee.Print("No se ha encontrado al cliente "+s);
+					}
+					
 					break;
 				case 4:
 					lee.Print("Introduzca el dni del cliente que quiera buscar");
-					Cliente c = new Cliente();
 					s=lee.String();
-					c=clientes.searchCliente(s);
-					lee.Print(c);
+					if(clientes.searchCliente(s)!=null) {
+						Cliente c = new Cliente();
+						c=clientes.searchCliente(s);
+						lee.Print(c);
+					}else {
+						lee.Print("No se ha encontrado al cliente "+s);
+					}
 					break;
 				case 5:
 					lee.show(clientes);
@@ -118,19 +145,32 @@ public class Submenus {
 				case 2:
 					lee.Print("Introduzca el id de la reserva a eliminar");
 					s=lee.String();
-					reservas.deleteReserva(s);
+					if(reservas.searchReserva(s)!=null) {
+						reservas.deleteReserva(s);
+						lee.Print("Reserva "+s+" eliminada");
+					}else {
+						lee.Print("No se ha podido eliminar la reserva "+s+" o no existe");
+					}
 					break;
 				case 3:
 					lee.Print("Introduzca el id de la reserva que quiera modificar");
 					s=lee.String();
-					Editmenu.Reservas(s, reservas);
+					if(reservas.searchReserva(s)!=null) {
+						Editmenu.Reservas(s, reservas);
+					}else {
+						lee.Print("No se ha podido encontrar la reserva "+s+" o no existe");
+					}
 					break;
 				case 4:
 					lee.Print("Introduzca el id de la reserva que quiera buscar");
 					s=lee.String();
-					Reserva r = new Reserva();
-					r=reservas.searchReserva(s);
-					lee.Print(r);
+					if(reservas.searchReserva(s)!=null) {
+						Reserva r = new Reserva();
+						r=reservas.searchReserva(s);
+						lee.Print(r);
+					}else {
+						lee.Print("No se ha podido encontrar la reserva "+s+" o no existe");
+					}
 					break;
 				case 5:
 					lee.show(reservas);
